fix: changed long to integer in month repo | update: created month service and implemented createrecord

+++ src/main/java/com/hishabi/api/repository/MonthRepository.java
@@ -8,6 +8,6 @@ import com.hishabi.api.entity.MonthEntity;
 @Repository
 public interface MonthRepository extends JpaRepository<MonthEntity, Long> {
 
-    boolean existsByUser_IdAndMonthAndYear(Long userId, Long month, Integer year);
+    boolean existsByUser_IdAndMonthAndYear(Long userId, Integer month, Integer year);
 
 }
